requirements.txt
# Meta-Agent Nova dependencies
openai>=1.0.0
langchain
crewai
docker
python-dotenv
grpcio>=1.58.0
